stress-futex: use long type for nanoseconds to match timespec struct
Avoid casting uint64_t to long

+++ stress-futex.c
@@ -38,7 +38,7 @@ static const stress_help_t help[] = {
  *  stress_futex_wait()
  *     exercise futex_wait and every 16th time futex_waitv
  */
-static int stress_futex_wait(uint32_t *futex, const int val, const uint64_t nsec)
+static int stress_futex_wait(uint32_t *futex, const int val, const long nsec)
 {
 	struct timespec t;
 
@@ -78,7 +78,7 @@ static int stress_futex_wait(uint32_t *futex, const int val, const uint64_t nsec
 	/* UNEXPECTED */
 #endif
 	t.tv_sec = 0;
-	t.tv_nsec = nsec;
+	t.tv_nsec = (long)nsec;
 
 	return shim_futex_wait(futex, val, &t);
 }
